+++ opentelemetry-proto/src/transform/common.rs
@@ -132,15 +132,16 @@ pub mod tonic {
     }
 }
 
-#[cfg(feature = "gen-protoc")]
+#[cfg(feature = "gen-grpcio")]
 pub mod grpcio {
-    use crate::proto::grpcio::common::{AnyValue, ArrayValue, InstrumentationScope, KeyValue};
+    use crate::proto::grpcio::common::v1::{
+        any_value, AnyValue, ArrayValue, InstrumentationScope, KeyValue,
+    };
     use opentelemetry_api::{Array, Value};
-    use opentelemetry_sdk::Resource;
-    use protobuf::RepeatedField;
-    #[cfg(feature = "logs")]
-    use protobuf::SingularPtrField;
     use std::borrow::Cow;
+
+    #[cfg(any(feature = "trace", feature = "logs"))]
+    use opentelemetry_sdk::Resource;
 
     impl From<opentelemetry_sdk::InstrumentationLibrary> for InstrumentationScope {
         fn from(library: opentelemetry_sdk::InstrumentationLibrary) -> Self {
@@ -153,73 +154,83 @@ pub mod grpcio {
         }
     }
 
+    impl From<&opentelemetry_sdk::InstrumentationLibrary> for InstrumentationScope {
+        fn from(library: &opentelemetry_sdk::InstrumentationLibrary) -> Self {
+            InstrumentationScope {
+                name: library.name.to_string(),
+                version: library
+                    .version
+                    .as_ref()
+                    .map(ToString::to_string)
+                    .unwrap_or_default(),
+                attributes: Attributes::from(library.attributes.clone()).0,
+                ..Default::default()
+            }
+        }
+    }
+
+    /// Wrapper type for Vec<[`KeyValue`](crate::proto::grpcio::common::v1::KeyValue)>
     #[derive(Default)]
-    pub struct Attributes(pub ::protobuf::RepeatedField<crate::proto::grpcio::common::KeyValue>);
+    pub struct Attributes(pub ::std::vec::Vec<crate::proto::grpcio::common::v1::KeyValue>);
 
     #[cfg(feature = "trace")]
     impl From<opentelemetry_sdk::trace::EvictedHashMap> for Attributes {
         fn from(attributes: opentelemetry_sdk::trace::EvictedHashMap) -> Self {
-            Attributes(RepeatedField::from_vec(
+            Attributes(
                 attributes
                     .into_iter()
-                    .map(|(key, value)| {
-                        let mut kv: KeyValue = KeyValue::new();
-                        kv.set_key(key.as_str().to_string());
-                        kv.set_value(value.into());
-                        kv
+                    .map(|(key, value)| KeyValue {
+                        key: key.as_str().to_string(),
+                        value: Some(value.into()),
                     })
                     .collect(),
-            ))
+            )
         }
     }
 
     impl From<Vec<opentelemetry_api::KeyValue>> for Attributes {
         fn from(kvs: Vec<opentelemetry_api::KeyValue>) -> Self {
-            Attributes(RepeatedField::from_vec(
+            Attributes(
                 kvs.into_iter()
-                    .map(|api_kv| {
-                        let mut kv: KeyValue = KeyValue::new();
-                        kv.set_key(api_kv.key.as_str().to_string());
-                        kv.set_value(api_kv.value.into());
-                        kv
+                    .map(|api_kv| KeyValue {
+                        key: api_kv.key.as_str().to_string(),
+                        value: Some(api_kv.value.into()),
                     })
                     .collect(),
-            ))
+            )
         }
     }
 
     #[cfg(feature = "logs")]
     impl<K: Into<String>, V: Into<AnyValue>> FromIterator<(K, V)> for Attributes {
         fn from_iter<T: IntoIterator<Item = (K, V)>>(iter: T) -> Self {
-            Attributes(RepeatedField::from_vec(
+            Attributes(
                 iter.into_iter()
                     .map(|(k, v)| KeyValue {
                         key: k.into(),
-                        value: SingularPtrField::some(v.into()),
-                        ..Default::default()
+                        value: Some(v.into()),
                     })
                     .collect(),
-            ))
+            )
         }
     }
 
     impl From<Value> for AnyValue {
         fn from(value: Value) -> Self {
-            let mut any_value = AnyValue::new();
-            match value {
-                Value::Bool(val) => any_value.set_bool_value(val),
-                Value::I64(val) => any_value.set_int_value(val),
-                Value::F64(val) => any_value.set_double_value(val),
-                Value::String(val) => any_value.set_string_value(val.to_string()),
-                Value::Array(array) => any_value.set_array_value(match array {
-                    Array::Bool(vals) => array_into_proto(vals),
-                    Array::I64(vals) => array_into_proto(vals),
-                    Array::F64(vals) => array_into_proto(vals),
-                    Array::String(vals) => array_into_proto(vals),
-                }),
-            };
-
-            any_value
+            AnyValue {
+                value: match value {
+                    Value::Bool(val) => Some(any_value::Value::BoolValue(val)),
+                    Value::I64(val) => Some(any_value::Value::IntValue(val)),
+                    Value::F64(val) => Some(any_value::Value::DoubleValue(val)),
+                    Value::String(val) => Some(any_value::Value::StringValue(val.to_string())),
+                    Value::Array(array) => Some(any_value::Value::ArrayValue(match array {
+                        Array::Bool(vals) => array_into_proto(vals),
+                        Array::I64(vals) => array_into_proto(vals),
+                        Array::F64(vals) => array_into_proto(vals),
+                        Array::String(vals) => array_into_proto(vals),
+                    })),
+                },
+            }
         }
     }
 
@@ -227,15 +238,12 @@ pub mod grpcio {
     where
         Value: From<T>,
     {
-        let values = RepeatedField::from_vec(
-            vals.into_iter()
-                .map(|val| AnyValue::from(Value::from(val)))
-                .collect(),
-        );
+        let values = vals
+            .into_iter()
+            .map(|val| AnyValue::from(Value::from(val)))
+            .collect();
 
-        let mut array_value = ArrayValue::new();
-        array_value.set_values(values);
-        array_value
+        ArrayValue { values }
     }
 
     #[cfg(any(feature = "trace", feature = "logs"))]